build: Refactor three.js imports of tests and remaining files.
