+++ lib/vpt/light/light-mesh.spec.ts
@@ -18,8 +18,8 @@
  */
 
 import { expect } from 'chai';
-import { SpotLight } from 'three';
 import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader';
+import { SpotLight } from 'three/src/lights/SpotLight';
 import { ThreeHelper } from '../../../test/three.helper';
 import { NodeBinaryReader } from '../../io/binary-reader.node';
 import { Table } from '../table/table';
