+++ lib/vpt/table/table.spec.ts
@@ -18,8 +18,8 @@
  */
 
 import { expect } from 'chai';
-import { Mesh } from 'three';
 import { GLTF } from 'three/examples/jsm/loaders/GLTFLoader';
+import { Mesh } from 'three/src/objects/Mesh';
 import { ThreeHelper } from '../../../test/three.helper';
 import { NodeBinaryReader } from '../../io/binary-reader.node';
 import { Table } from './table';
